camlistore-configure: replace script with a go program
/usr/local/bin/camlistore-configure on the final image is now a Go
program compiled from patches/usr/local/src/camlistore-configure.go during
the building of the image.

The reasons for that change are twofold:

1) Instead of calling camlistored to generate the config file, and
having to grossly kill it afterwards, this code (using some
camlistore.org imports) now just does this generation, without
having to call camlistored.

2) go flags are self-describing, so it will be easier for users to get
the usage of this program than with a shell script.

Tested with make shell, but not yet as a new image (which is the only
true way to test if the systemctl calls were successful).
Will do Tomorrow.